Move helpers and selectors into same subfolder as Configuration Management integration tests
